+++ lib/weak/map/abstract_strong_keys.rb
@@ -5,7 +5,7 @@
 # This software may be modified and distributed under the terms
 # of the MIT license. See the LICENSE.txt file for details.
 
-require "weak/map/deletable"
+require_relative "deletable"
 
 ##
 module Weak