+++ lib/weak/map/weak_keys.rb
@@ -5,8 +5,8 @@
 # This software may be modified and distributed under the terms
 # of the MIT license. See the LICENSE.txt file for details.
 
-require "weak/map/deletable"
-require "weak/undefined"
+require_relative "deletable"
+require_relative "../undefined"
 
 ##
 module Weak