+++ CHANGELOG.md
@@ -8,10 +8,10 @@
 ## [0.2.1] - 2025-12-27
 
 - Fix typos in code documentation
-- Ignore some unnecessary methods defined on some `Set` implementations in `set_spec`
+- Ignore some unnecessary methods defined on some `Weak::Set` implementations in `set_spec`
 - Run specs on JRuby 10 in Github Actions
 - Retry TruffleRuby rspec runs on Github Actions to avoid random failures due to flakey GC.
-- Extract UNDEFINED to its own file and require it where used.
+- Extract `UNDEFINED` to its own file and require it where used.
 - Add more details about the gem version in `Weak::Version`
 - Handle object cycles in `pretty_print`.