+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix `Weak::Map#store` method alias to `Weak::Map#[]=`. Previously, it was errouneously aliased to `Weak::Map#[]`.
 - Fix typos in code documentation
 - Use require_relative instead of require for all gem files
+- Add addititional specs for `Weak::Map`
 
 ## [0.2.1] - 2025-12-27