+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## [Unreleased]
 
-- Adapt `Weak::Set#inspect` output to more resemble the output of `Set#inspect` in Ruby 4.0
+### Bug Fixes
+
 - Fix `Weak::Map#store` method alias to `Weak::Map#[]=`. Previously, it was errouneously aliased to `Weak::Map#[]`.
+
+### Improvements
+
+- Adapt `Weak::Set#inspect` output to more resemble the output of `Set#inspect` in Ruby 4.0
 - Fix typos in code documentation
-- Use require_relative instead of require for all gem files
+- Use `require_relative` instead of require for all gem files
 - Add addititional specs for `Weak::Map`
 
 ## [0.2.1] - 2025-12-27